Split SearchScreen into a search package
SearchScreen.kt was 1688 lines: the ViewModel, the screen, five distinct
empty/placeholder states, the curated-list card, and ten previews.
Continues the ui/word/ shape applied to Favorites and Settings.

  ui/search/SearchViewModel.kt        UiState and ViewModel        443
  ui/search/SearchScreen.kt           Screen, Content, result card 411
  ui/search/SearchEmptyStates.kt      the five empty states        515
  ui/search/ListCard.kt               curated-list card            219
  ui/search/preview/SearchPreviews.kt the ten previews             483

The list colour helpers do not go into ui/search. listHue and
vibrantColorsForList assign a stable brand colour per list id, and
ListDetailScreen uses vibrantColorsForList too, so filing them under the
search screen would have made ListDetail depend on Search for a value
that is really a property of a word list. They move to
ui/components/ListColors.kt, which both import. mixColors stays private
in ListCard.kt, where it is the only caller.

Pure move. Diffing the moved bodies against the original shows the only
changes are four widenings forced by the new file boundaries:

  EmptySearchState, NoDictionaryState, NoResultsState  private -> internal
  ListCard                                             private -> internal

ReadSection, WordChip, SuggestionCard and ScrollableNoDictionaryContainer
keep private: each is used only by the state that now shares its file.
