+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/ui/ListDetailScreen.kt
@@ -46,6 +46,7 @@ import com.slovy.slovymovyapp.data.Language
 import com.slovy.slovymovyapp.data.favorites.FavoritesRepository
 import com.slovy.slovymovyapp.data.favorites.NewFavorite
 import com.slovy.slovymovyapp.ui.favorites.toFavoriteSenseLoadError
+import com.slovy.slovymovyapp.ui.components.vibrantColorsForList
 import com.slovy.slovymovyapp.data.lists.ListsService
 import com.slovy.slovymovyapp.data.lists.WordList
 import com.slovy.slovymovyapp.data.lists.WordListSense
